Update clust_homo.py
I didn't delete the run_usearch function in case you wanted to keep it, so I just commented it out. But basically, it seemed easier to do uclust this way by letting the user decide which option(s) they want, and then it will run usearch based on what the user wanted. So I have pseudocode too for that (indicated by the ##)